Fix the timing
The code had various timing issues related to things like using milliseconds for some purposes, and FreeRTOS ticks for others.

Now the only instances of `millis()` is in the code to measure actual drift, because I can't use `xTaskGetTickCount()` because it doesn't work inside an ISR.

I was also doing some questionable time arithmetic, and it turned out that I needed to ensure that `RunMotors()` took some time to complete even if the sequence to run was silent.

I'm biasing the client's time drift by -15ms, because I have determined experimentally that the actual drift oscillates around 15±10 ms. Probably not a coincidence taking into account that the BLE scan interval is set at 10ms.

This should work for now, but maybe I can improve on this later by using ESP-NOW instead of BLE for synchronization.